POC: Started example for logs endpoint
